04/01/2016 14:22

+++ src/java/models/Movie.java
@@ -18,7 +18,7 @@ public class Movie implements DBEntity {
     protected int id;
     protected String name;
     protected Date release_date;
-    protected int movie_length;
+    protected double movie_length;
     protected MovieCategory category;
     protected String plot;
     protected String poster;
@@ -70,11 +70,11 @@ public class Movie implements DBEntity {
         this.release_date = release_date;
     }
 
-    public int getMovie_length() {
+    public double getMovie_length() {
         return movie_length;
     }
 
-    public void setMovie_length(int movie_length) {
+    public void setMovie_length(double movie_length) {
         this.movie_length = movie_length;
     }
